refactor: modularize admin char shop
