+++ admin.js
@@ -6,7 +6,6 @@ const fs = require('node:fs');
 const path = require('node:path');
 const clientManager = require('./clientManager');
 
-const mapOptions = ["Name", "About", "Channels", "Image", "Emoji"]
 
 class Admin {
 
@@ -295,275 +294,6 @@ When selected grants the:
     return "Shire " + shireName + " has been added with resource " + resource;
   }
 
-  //addMap adds a new map to data, should be similar to addRecipe NOT to addShire
-  static async addMap(mapName, guild, mapType = "map") {
-    // Load the maps collection
-    let data = await dbm.loadFile('keys', mapType + 's');
-    let mapNames = Object.keys(data);
-    let i = 1;
-    let newMapName = mapName;
-  
-    // Ensure the map name is unique by appending a number if necessary
-    while (mapNames.includes(newMapName)) {
-      newMapName = mapName + " " + i;
-      i++;
-    }
-  
-    // Create a new map object with all fields blank
-    let mapData = {
-      "mapOptions": mapOptions.reduce((acc, option) => {
-        acc[option] = "";
-        return acc;
-      }, {}),
-    };
-    mapData.mapOptions.Name = newMapName;
-    mapData.mapOptions.Emoji = ":map:";
-
-    data[newMapName] = mapData;
-  
-    // Save the new map to the maps collection
-    await dbm.saveFile('keys', mapType + 's', data); 
-  
-    return newMapName;
-  }
-
-  static async editMapMenu(mapName, tag, mapType = "map") {
-    // Load the map data
-    let mapData = await dbm.loadFile('keys', mapType + 's');
-  
-    if (mapData[mapName] == undefined) {
-      for (let key in mapData) {
-        if (key.toLowerCase() == mapName.toLowerCase()) {
-          mapName = key;
-          break;
-        }
-      }
-      if (mapData[mapName] == undefined) {
-        return "Map not found!";
-      }
-    }
-  
-    mapData = mapData[mapName];
-  
-    let userData = await dbm.loadFile('characters', tag);
-    if (!userData.editingFields) {
-      userData.editingFields = {};
-    }
-    userData.editingFields["Map Edited"] = mapName;
-    userData.editingFields["Map Type Edited"] = mapType;
-    await dbm.saveFile('characters', tag, userData);
-  
-    // Construct the edit menu embed
-    const embed = new EmbedBuilder()
-      .setTitle("**" + mapName + "**")
-      .setDescription('Edit the fields using the command /editmapfield <field number> <new value>');
-    
-    let emoji = mapData.mapOptions.Emoji;
-
-    let valueString = mapOptions.map((option, index) => {
-      let value = mapData.mapOptions[option];
-      
-      // If index is 1 and value is too long, replace it
-      if (index === 1 && value.length > 50) {
-          value = "Too long! Must view in /editembedabout";
-      }
-  
-      return `\`[${index + 1}] ${option}:\` ` + value;
-  }).join('\n');
-  
-  
-    // Add fields for Map Options
-    embed.addFields({ name: emoji + ' Embed Options', value: valueString });
-    embed.setFooter({ text: 'Page 1 of 1, Map Options' });
-  
-    // Return the embed
-    return embed;
-  }
-
-  static async removeMap(mapName, mapType = "map") {
-    // Load the maps collection
-    let data = await dbm.loadFile('keys', mapType + 's');
-    if (data[mapName] == undefined) {
-      return "Map not found! Must match the exact name of the map, case sensitive."
-    }
-  
-    // Delete the map from the collection
-    delete data[mapName];
-  
-    // Save the updated collection
-    await dbm.saveFile('keys', mapType + 's', data);
-  
-    return mapName + ' has been deleted';
-  }
-
-  static async editMapAbout(interaction) {
-    let infoSection = interaction.customId.replace("editmapaboutmodal", "");
-    let mapNameNoSpaces = infoSection.split("||")[0];
-    let mapType = infoSection.split("||")[1];
-    let maps = await dbm.loadFile('keys', mapType + 's');
-    let mapName = Object.keys(maps).find(key => key.replace(/ /g, '').toLowerCase() == mapNameNoSpaces);
-    if (mapName == undefined) {
-      interaction.reply("Map not found!");
-      return;
-    }
-    console.log(interaction)
-    let about = interaction.fields.getTextInputValue('mapabout');
-
-    maps[mapName].mapOptions.About = about;
-    await dbm.saveFile('keys', mapType + 's', maps);
-    interaction.reply("About section has been updated");
-  }
-  
-  static async editMapField(charTag, field, value) {
-    // Load the maps collection
-    let charData = await dbm.loadFile('characters', charTag);
-    if (!charData.editingFields || !charData.editingFields["Map Edited"] || !charData.editingFields["Map Type Edited"]) {
-      return "You must use /editmapmenu first to select a map to edit";
-    }
-    let data = await dbm.loadFile('keys', charData.editingFields["Map Type Edited"] + 's');
-    let mapName = charData.editingFields["Map Edited"];
-    let mapType = charData.editingFields["Map Type Edited"];
-    if (data[mapName] == undefined) {
-      return "Map not found! Must match the exact name of the map, case sensitive."
-    }
-
-    //Field is a number, so we need to convert it to the actual field name
-    field = mapOptions[field - 1];
-    if (field == undefined) {
-      return "Field not found! Must be a number between 1 and " + mapOptions.length
-    }
-    //If field is 3, make sure that its in the format <#channel1> <#channel2> etc.
-    if (field == "Channels") {
-      let unformattedChannels = value.split("<#");
-      let channelString = "";
-      //Make sure every channel has a > at the end. For any channel that has one, add it formatted with a comma to the channelString
-      for (let i = 0; i < unformattedChannels.length; i++) {
-        let channel = unformattedChannels[i];
-        if (channel.includes(">")) {
-          channelString += "<#" + channel.split(">")[0] + ">";
-          if (i < unformattedChannels.length - 1) {
-            channelString += ", ";
-          }
-        }
-      }
-      //If final character is a comma, remove it
-      if (channelString[channelString.length - 1] == ",") {
-        channelString = channelString.slice(0, -1);
-      }
-      value = channelString;
-    }
-
-    //If field is name, make sure it is unique and than update the map name
-    if (field == "Name") {
-      if (data[value] != undefined) {
-        return "Map name already exists!";
-      }
-      data[value] = data[mapName];
-      delete data[mapName];
-      mapName = value;
-    }
-  
-    // Update the field with the new value
-    data[mapName].mapOptions[field] = value;
-  
-    // Save the updated collection
-    await dbm.saveFile('keys', mapType + 's', data);
-
-    if (value == "") {
-      return 'Field ' + field + ' has been removed';
-    }
-  
-    return 'Field ' + field + ' has been updated to ' + value;
-  }
-
-  static async allMaps() {
-    let maps = await dbm.loadFile("keys", "maps");
-    //Create an embed with the title ":map: All Maps", and than the description is a list of all maps in the form <Emoji> **<MapName>**
-    let mapNames = Object.keys(maps).map(key => maps[key].mapOptions.Emoji + " **" + key + "**").join("\n");
-    let embed = new EmbedBuilder()
-      .setTitle("All Maps")
-      .setDescription(mapNames);
-
-    return embed;
-  }
-
-  static async allGuides() {
-    let guides = await dbm.loadFile("keys", "guides");
-    let mapNames = Object.keys(guides).map(key => guides[key].mapOptions.Emoji + " **" + key + "**").join("\n");
-    let embed = new EmbedBuilder()
-      .setTitle("All Guides")
-      .setDescription(mapNames);
-
-    return embed;
-  }
-
-  static async allLores() {
-    let lores = await dbm.loadFile("keys", "lores");
-    let mapNames = Object.keys(lores).map(key => lores[key].mapOptions.Emoji + " **" + key + "**").join("\n");
-    //Iterate over each value, log emoji + " " + key
-    for (const key in lores) {
-      console.log(lores[key]);
-    }
-    let embed = new EmbedBuilder()
-      .setTitle("All Lores")
-      .setDescription(mapNames);
-
-    return embed;
-  }
-
-  static async allRanks() {
-    let ranks = await dbm.loadFile("keys", "ranks");
-    let mapNames = Object.keys(ranks).map(key => ranks[key].mapOptions.Emoji + " **" + key + "**").join("\n");
-    let embed = new EmbedBuilder()
-      .setTitle("All Ranks")
-      .setDescription(mapNames);
-
-    return embed;
-  }
-
-  static async map(mapName, channelId, type = "map") {
-    let maps = await dbm.loadFile("keys", type + "s");
-    let map = maps[mapName];
-    while (map == undefined) {
-      for (const key in maps) {
-        if (key.toLowerCase().replace(/ /g, '') == mapName.toLowerCase().replace(/ /g, '')) {
-          map = maps[key];
-          mapName = key;
-        }
-      }
-      if (map == undefined) {
-        return "Map not found!";
-      }
-    }
-
-    if (map.mapOptions.Channels != "") {
-      //Check if channelId is a channel in the map
-      let channels = map.mapOptions.Channels.split("<#");
-      let channelIDs = channels.map(channel => channel.split(">")[0]);
-      if (!channelIDs.includes(channelId)) {
-        return "This map is not available in this channel!";
-      }
-    }
-
-    let embed;
-    if (map.mapOptions.Emoji != "") {
-      embed = new EmbedBuilder().setTitle(map.mapOptions.Emoji + " " + mapName);
-    } else {
-      embed = new EmbedBuilder().setTitle(mapName);
-    }
-    if (map.mapOptions.About != "") {
-      embed.setDescription(map.mapOptions.About);
-    }
-    if (map.mapOptions.Image != "") {
-      embed.setImage(map.mapOptions.Image);
-    }
-    //Typename
-    let typeName = type.charAt(0).toUpperCase() + type.slice(1);
-
-    embed.setFooter({ text: typeName + ' of ' + mapName });
-  
-    return embed;
-  }
   
 
   static async selectShire(interaction) {
