+++ commands/shopCommands/editrecipefield.js
@@ -1,11 +1,11 @@
 const { SlashCommandBuilder } = require('discord.js');
-const shop = require('../../shop'); // Importing the database manager
+const crafting = require('../../shop/crafting'); // Importing the database manager
 
 ///editfield <field number> <new value>
 module.exports = {
 	data: new SlashCommandBuilder()
         .setName('editrecipefield')
-        .setDescription('Edit a field of an recipe in the shop. Use /editrecipemenu first to see the fields of an recipe')
+        .setDescription('Edit a field of an recipe in the crafting. Use /editrecipemenu first to see the fields of an recipe')
         .setDefaultMemberPermissions(0)
         .addIntegerOption((option) =>
             option.setName('fieldnumber')
@@ -21,7 +21,7 @@ module.exports = {
         const fieldNumber = interaction.options.getInteger('fieldnumber');
         const newValue = interaction.options.getString('newvalue');
 
-        let reply = await shop.editRecipeField(interaction.user.tag, fieldNumber, newValue);
+        let reply = await crafting.editRecipeField(interaction.user.tag, fieldNumber, newValue);
         await interaction.reply(reply);
     }
 };
