+++ shop.js
@@ -15,13 +15,6 @@ class shop {
       'Take Item', 'Take Item 2', 'Take Item 3', 'Take Item 4', 'Take Item 5',
       'Change Legitimacy (#)', 'Change Prestige (#)', 'Change Martial (#)', 'Change Intrigue (#)', 'Change Devotion (#)', 'Revive (Y/N)', 'Durability (#)'
     ];
-  static recipeOptions = [
-      'Name', 'Icon', 'Show Image', 'Show Message',
-      'Ingredient 1', 'Ingredient 2', 'Ingredient 3', 'Ingredient 4', 'Ingredient 5', 'Ingredient 6', 'Ingredient 7', 'Ingredient 8', 'Ingredient 9', 'Ingredient 10',
-      'Result 1', 'Result 2', 'Result 3', 'Result 4', 'Result 5',
-      'Craft Time in Hours (#)', 'Need None Of Roles', 'Need All Of Roles', 'Need Any Of Roles', 'Is Public (Y/N)'
-    ];
-
   // Function to find an item by name in the shop
   //THIS IS INEFFICIENT BECAUSE IT MEANS CALLING IT MEANS TWO CALLS TO THE DATABASE- FIX LATER
   static async findItemName(itemName, data) {
@@ -85,146 +78,6 @@ class shop {
     await dbm.saveFile('shop', itemName, itemData);
   }
 
-  static async addRecipe(recipeName) {
-    //Go into the recipes file. First check if another copy of this recipe exists. If it does, add a space and number to recipe name and check again
-    let data = await dbm.loadCollection('recipes');
-    let shopData = await dbm.loadCollection('shop');
-    let recipeNames = Object.keys(data);
-    let i = 1;
-    let newRecipeName = recipeName;
-    while (recipeNames.includes(newRecipeName)) {
-      newRecipeName = recipeName + " " + i;
-      i++;
-    }
-    //Create a new recipe object with all fields blank
-    let recipeData = {
-      "recipeOptions": this.recipeOptions.reduce((acc, option) => {
-        acc[option] = "";
-        return acc;
-      }
-      , {}),
-    };
-
-    //Set option "Is Public (Y/N)" to No
-
-    recipeData.recipeOptions["Is Public (Y/N)"] = "No";
-    let itemName = await this.findItemName(newRecipeName, shopData);
-    if (itemName != "ERROR") {
-      let itemData = shopData[itemName];
-      newRecipeName = itemName;
-      recipeData.recipeOptions.Name = itemName;
-      recipeData.recipeOptions.Icon = itemData.infoOptions.Icon;
-      recipeData.recipeOptions["Result 1"] = "1 " + itemName;
-    } else {
-      recipeData.recipeOptions.Name = newRecipeName;
-      recipeData.recipeOptions.Icon = ":hammer:";
-    }
-    recipeData.recipeOptions["Craft Time in Hours (#)"] = 1;
-    await dbm.saveFile('recipes', newRecipeName, recipeData); 
-
-    return newRecipeName;
-  }
-
-  static async recipesEmbed(isPublic, page) {
-    const itemsPerPage = 1000; // Number of recipes per page
-    let data = await dbm.loadCollection('recipes');
-    let shopData = await dbm.loadCollection('shop');
-    let publicRecipes = [];
-    let privateRecipes = [];
-  
-    //Loop through data 
-    for (let [key, value] of Object.entries(data)) {
-      if (value.recipeOptions["Is Public (Y/N)"] == "Yes") {
-        publicRecipes.push(value);
-      } else {
-        privateRecipes.push(value);
-      }
-    }
-
-
-    let recipesToShow = isPublic ? publicRecipes : publicRecipes.concat(privateRecipes);
-
-    let categorizedRecipes = {};
-    let itemNamesLower = Object.keys(shopData).map(name => name.toLowerCase());
-    for (let recipe of recipesToShow) {
-      let category = "Uncategorized";
-      if (shopData[recipe.recipeOptions.Name]) {
-        category = shopData[recipe.recipeOptions.Name].infoOptions.Category || category;
-      } else {
-        // Search for comparison lower case
-        let recipeName = recipe.recipeOptions.Name.toLowerCase();
-        let index = itemNamesLower.indexOf(recipeName);
-        if (index != -1) {
-          category = shopData[Object.keys(shopData)[index]].infoOptions.Category || category
-        }
-      }
-      if (category == "") {
-        category = "Uncategorized";
-      }
-      if (!categorizedRecipes[category]) {
-        categorizedRecipes[category] = [];
-      }
-      categorizedRecipes[category].push(recipe);
-    }
-  
-    // Pagination calculation
-    const pageStart = (page - 1) * itemsPerPage;
-    const pageEnd = pageStart + itemsPerPage;
-    const totalPages = Math.ceil(recipesToShow.length / itemsPerPage);
-  
-    let returnEmbed = new Discord.EmbedBuilder()
-      .setTitle(':hammer: Recipes')
-      .setColor(0x36393e)
-      .setFooter({ text: `Page ${page} of ${totalPages}` });
-
-    let descriptionText = '';
-
-    const addRecipesToDescription = (recipes, startIndex, endIndex, category, isPublic) => {
-      if (recipes.length > 0) {
-        descriptionText += `**${category}**\n`;
-        for (let i = startIndex; i < endIndex && i < recipes.length; i++) {
-          let recipeName = recipes[i].recipeOptions.Name;
-          if (!isPublic && privateRecipes.includes(recipes[i])) {
-            recipeName += " :warning:";
-          }
-          descriptionText += (recipes[i].recipeOptions.Icon ? recipes[i].recipeOptions.Icon + " " : ":hammer: ") + recipeName + "\n";
-        }
-      }
-    };
-
-    let count = 0;
-
-    for (let [category, recipes] of Object.entries(categorizedRecipes)) {
-      if (count >= pageStart && count < pageEnd) {
-        let startIndex = Math.max(pageStart - count, 0);
-        let endIndex = Math.min(pageEnd - count, recipes.length);
-        addRecipesToDescription(recipes, startIndex, endIndex, category, isPublic);
-      }
-      count += recipes.length;
-    }
-
-    if (descriptionText === '') {
-      descriptionText = 'No recipes found!';
-    }
-
-    returnEmbed.setDescription(descriptionText);
-  
-    // Buttons for navigation
-    const prevButton = new ButtonBuilder()
-      .setCustomId('prev_page')
-      .setLabel('<')
-      .setStyle(ButtonStyle.Secondary)
-      .setDisabled(page === 1);
-  
-    const nextButton = new ButtonBuilder()
-      .setCustomId('next_page')
-      .setLabel('>')
-      .setStyle(ButtonStyle.Secondary)
-      .setDisabled(page === totalPages);
-  
-    let actionRow = new ActionRowBuilder().addComponents(prevButton, nextButton);
-  
-    return [returnEmbed, actionRow];
   }
   
 
@@ -1061,46 +914,6 @@ class shop {
     return [embed, rows];
   }
 
-  static async editRecipeMenu(recipeName, tag) {
-    // Load the recipe data
-    let recipeData = await dbm.loadCollection('recipes', recipeName);
-
-    if (recipeData[recipeName] == undefined) {
-      for (let key in recipeData) {
-        if (key.toLowerCase() == recipeName.toLowerCase()) {
-          recipeName = key;
-          break;
-        }
-      }
-      if (recipeData[recipeName] == undefined) {
-        return "Recipe not found!";
-      }
-    }
-
-    recipeData = recipeData[recipeName];
-
-    let userData = await dbm.loadCollection('characters');
-    if (!userData[tag].editingFields) {
-      userData[tag].editingFields = {};
-    }
-    userData[tag].editingFields["Recipe Edited"] = recipeName;
-    await dbm.saveCollection('characters', userData);
-
-    const recipeOptions = this.recipeOptions;
-
-    // Construct the edit menu embed
-    const embed = new Discord.EmbedBuilder()
-      .setTitle("**" + recipeName + "**")
-      .setDescription('Edit the fields using the command /editrecipefield <field number> <new value>');
-
-    // Add fields for Recipe Options
-    embed.addFields({ name: '📜 Recipe Options', value: recipeOptions.map((option, index) => `\`[${index + 1}] ${option}:\` ` + recipeData.recipeOptions[option]).join('\n') });
-    embed.setFooter({text : 'Page 1 of 1, Recipe Options'});
-
-    //Return an array including the embed
-    return embed;
-  }
-
   static async editItemField(userTag, fieldNumber, newValue) {
     let shopData = await dbm.loadCollection('shop');
     // Load user data
@@ -1255,204 +1068,6 @@ class shop {
     }
     return `Field ${fieldName} updated to ${newValue} for item ${itemName}`;
   }
-
-  static async editRecipeField(userTag, fieldNumber, newValue) {
-    // Load user data
-    let userData = await dbm.loadCollection('characters');
-    let recipeName;
-    if (!userData[userTag].editingFields["Recipe Edited"]) {
-      return "You are not currently editing any recipes!";
-    } else {
-      recipeName = userData[userTag].editingFields["Recipe Edited"];
-    }
-
-    // Load the recipe data
-    let recipeData = await dbm.loadFile('recipes', recipeName);
-
-    let shopData = await dbm.loadCollection('shop');
-
-    const recipeOptions = this.recipeOptions;
-
-    // Determine which category the field number belongs to
-    let category;
-    if (fieldNumber >= 1 && fieldNumber <= recipeOptions.length) {
-      category = 'recipeOptions';
-    } else {
-      return "Invalid field number!";
-    }
-
-    // Get the field name
-    let fieldName;
-    switch (category) {
-      case 'recipeOptions':
-        fieldName = recipeOptions[fieldNumber - 1];
-        break;
-    }
-
-    let nullValue = false;
-    if (newValue == null) {
-      newValue = "";
-      nullValue = true;
-    }
-
-    if (nullValue) {
-      recipeData[category][fieldName] = newValue;
-      await dbm.saveFile('recipes', recipeName, recipeData);
-      return `Field ${fieldName} reset to blank for recipe ${recipeName}`;
-    }
-
-    //If category contains #, convert newValue to number- if it's not a number, return an error
-    if (fieldName.includes("#")) {
-      if (fieldName == "Craft Time in Hours (#)") {
-        //Check if its just a single number at first
-        let num = parseInt(newValue);
-        if (isNaN(num)) {
-          //Check if its in the format "<#> <Unit>"
-          let splitString = newValue.split(" ");
-          num = parseInt(splitString[0]);
-          if (isNaN(num)) {
-            return "Invalid value for number! This should be given in the form <Number> <Unit>";
-          }
-          //Check if unit is valid- should start with h, d, w, or m though can be upper or lower case
-          let unit = splitString[1].toLowerCase();
-          unit = unit.charAt(0);
-          if (unit != "h" && unit != "d" && unit != "w" && unit != "m") {
-            return "Invalid value for unit! This should be given in the form <Number> <Unit>";
-          } else {
-            //Convert to hours
-            switch (unit) {
-              case "d":
-                newValue = num * 24;
-                break;
-              case "w":
-                newValue = num * 24 * 7;
-                break;
-              case "m":
-                newValue = num * 24 * 30;
-                break;
-              case "h":
-                newValue = num;
-                break;
-              default:
-                return "Invalid value for unit! This should be given in the form <Number> <Unit>";
-            }
-          }
-        } else {
-          newValue = num;
-        }
-      } else { 
-        let num = parseInt(newValue);
-        if (isNaN(num)) {
-          return "Invalid value for a number field!";
-        }
-        newValue = num;
-      }
-    }
-
-
-    if (fieldName.includes("Y/N")) {
-      if (newValue.toLowerCase() == "y" || newValue.toLowerCase() == "yes" || newValue.toLowerCase() == "true") {
-        newValue = "Yes";
-      } else if (newValue.toLowerCase() == "n" || newValue.toLowerCase() == "no" || newValue.toLowerCase() == "false") {
-        newValue = "No";
-      } else {
-        return "Invalid value for a Y/N field!";
-      }
-    }
-
-    if (fieldName.includes("Ingredient") || fieldName.includes("Result")) {
-      //Should be in the form NUMBER ITEM NAME
-      let splitString = newValue.split(" ");
-      let num = parseInt(splitString[0]);
-      if (isNaN(num)) {
-        return "Invalid value for number! This should be given in the form <Number> <Item Name>";
-      }
-      //Check if item name is valid
-      let itemName = splitString.slice(1).join(" ");
-      let foundItemName = await this.findItemName(itemName, shopData);
-      if (foundItemName == "ERROR") {
-        return "Invalid value for item name! This should be given in the form <Number> <Item Name>";
-      } else {
-        newValue = num + " " + foundItemName;
-      } 
-    }
-
-    console.log(recipeData)
-
-    // Update the recipe data
-    recipeData[category][fieldName] = newValue;
-
-    // If there is now only one result (Result 1), and no other recipes exist with the name of that result, change the recipe name and icon to that result
-    if (fieldName == "Result 1" && recipeName.includes("New Recipe") && recipeData.recipeOptions["Result 2"] == "" && recipeData.recipeOptions["Result 3"] == "" && recipeData.recipeOptions["Result 4"] == "" && recipeData.recipeOptions["Result 5"] == "") {
-      let result = newValue.split(" ").slice(1).join(" ");
-
-      if (!recipeData[result]) {
-        let data = await dbm.loadCollection('shop');
-        if (data[result]) {
-          recipeData.recipeOptions["Name"] = result;
-          recipeData.recipeOptions["Icon"] = data[result].infoOptions.Icon;
-          
-          //Save new recipe
-          await dbm.saveFile('recipes', result, recipeData);
-          //Delete old recipe
-          await dbm.docDelete('recipes', recipeName);
-
-          //Return
-          return `This is now a recipe to craft ${result}, name and icon have been changed accordingly.`;
-        }   
-      }
-    }
-
-    // If the recipe name has changed, save the new recipe and delete the old one
-    if (fieldName == "Name") {
-      //Save new recipe
-      await dbm.saveFile('recipes', newValue, recipeData);
-      //Delete old recipe
-      await dbm.docDelete('recipes', recipeName);
-
-      //Change the recipe name in the user's editingFields
-      userData[userTag].editingFields["Recipe Edited"] = newValue;
-
-      //For each character in the userdata, in their cooldowns.craftSlots fields, replace the old name with the new name if it exists
-      let characters = Object.keys(userData);
-      for (let i = 0; i < characters.length; i++) {
-        if (userData[characters[i]].cooldowns && userData[characters[i]].cooldowns.craftSlots && characters[i] == "thegreatferret") {
-          let slots = userData[characters[i]].cooldowns.craftSlots;
-          console.log(slots);
-          //Slots is a json, not an array
-          let slotsKeys = Object.keys(slots);
-          console.log(slotsKeys);
-          for (let j = 0; j < slotsKeys.length; j++) {
-            let key = slotsKeys[j];
-            //Will either be recipeName or REPEAT_1_recipeName (repeat 2, etc), but you can't just check for inclusion because REPEAT_1_Woolen Tunic will include Wool
-            if (key == recipeName) {
-              slots[newValue] = slots[key];
-              delete slots[key];
-            }
-            //Check if first 7 characters are REPEAT_ and if its followed by a number and underscore
-            if (key.slice(0, 7) == "REPEAT_" && !isNaN(key.slice(7, 8)) && key.slice(8, 9) == "_" && key.slice(9) == recipeName) {
-              let newKey = "REPEAT_" + key.slice(7, 8) + "_" + newValue;
-              slots[newKey] = slots[key];
-              delete slots[key];
-            }
-          }
-          console.log(userData[characters[i]].cooldowns.craftSlots);
-          userData[characters[i]].cooldowns.craftSlots = slots;
-          console.log(userData[characters[i]].cooldowns.craftSlots);
-        }
-      }
-
-      await dbm.saveCollection('characters', userData);
-
-      return `Recipe name changed to ${newValue}`;
-    } else {
-      // Save the updated recipe data
-      await dbm.saveFile('recipes', recipeName, recipeData);
-    }
-    
-    
-    return `Field ${fieldName} updated to ${newValue} for recipe ${recipeName}`;
-  } 
 
   static async buyItem(itemName, charID, numToBuy, channelId) {
     let shopData = await dbm.loadCollection('shop');
